deploy script and addresses
