+++ contract-addresses/mainnetAddresses.json
@@ -39,6 +39,13 @@
     "creamFeiAddress" : { "artifact" : "CErc20Delegator", "address": "0x8C3B7a4320ba70f8239F83770c4015B5bc4e6F91" },
     "poolPartyFeiAddress" : { "artifact" : "CErc20Delegator", "address": "0x17b1A2E012cC4C31f83B90FF11d3942857664efc" },
     "indexCoopFusePoolFeiAddress" : { "artifact" : "CErc20Delegator", "address" : "0x04281F6715Dea6A8EbBCE143D86ea506FF326531" },
+    "rariPool8FeiPCVDepositAddress": { "artifact" : "ERC20CompoundPCVDeposit", "address": "0x37349d9cc523D28e6aBFC03fc5F44879bC8BfFD9"},
+    "rariPool6FeiPCVDepositAddress": { "artifact" : "ERC20CompoundPCVDeposit", "address": "0xB51f09B6F103D697dc5d64DC904Ad6a2Dad39987"},
+    "rariPool7FeiPCVDepositAddress": { "artifact" : "ERC20CompoundPCVDeposit", "address": "0x74B235Fef146cDB5BE0D3786a9f3774674b3615E"},
+    "rariPool24FeiPCVDepositAddress": { "artifact" : "ERC20CompoundPCVDeposit", "address": "0x1434F99EDB2bD03DECCCFe21288767b8324B7403"},
+    "creamFeiPCVDepositAddress" : { "artifact" : "ERC20CompoundPCVDeposit", "address": "0x243C601CC5DaA3Ac250B14509804188347bd2aFB" },
+    "poolPartyFeiPCVDepositAddress" : { "artifact" : "ERC20CompoundPCVDeposit", "address": "0x5A8CB4556e5D5935Af06beab8292905f48131479" },
+    "indexCoopFusePoolFeiPCVDepositAddress" : { "artifact" : "ERC20CompoundPCVDeposit", "address" : "0xD6960adba53212bBE96E54a7AFeDA2066437D000" },
     "bentoBoxAddress" : { "artifact" : "IMasterContractManager", "address" : "0xF5BCE5077908a1b7370B9ae04AdC565EBd643966" },
     "masterKashiAddress" : { "address" : "0x2cba6ab6574646badc84f0544d05059e57a5dc42" },
     "kashiFeiTribeAddress" : { "artifact" : "IKashiPair", "address": "0x18c9584d9ce56a0f62f73f630f180d5278c873b7" },
